Make deployd prioritise bouncing services on start
When deployd starts it must evaluate every service with a run of setup
marathon job to ensure it hasn't "missed" an event (e.g. autoscaling or
deployment). However, when we start we can prioritise services that we
know are deploying becuase their app ID has changed. This way if there
is a leader election or restart of deployd: bouncing/deploying services
will be at the front of the queue when it starts up.
