+++ paasta_tools/deployd/common.py
@@ -8,6 +8,13 @@ from threading import Thread
 
 from six.moves.queue import Queue
 
+from paasta_tools.marathon_tools import DEFAULT_SOA_DIR
+from paasta_tools.marathon_tools import get_marathon_client
+from paasta_tools.marathon_tools import list_all_marathon_app_ids
+from paasta_tools.marathon_tools import load_marathon_config
+from paasta_tools.marathon_tools import load_marathon_service_config_no_cache
+from paasta_tools.utils import NoDockerImageError
+
 BounceTimers = namedtuple('BounceTimers', ['processed_by_worker', 'setup_marathon', 'bounce_length'])
 ServiceInstance = namedtuple('ServiceInstance', ['service',
                                                  'instance',
@@ -62,3 +69,27 @@ def rate_limit_instances(instances, number_per_minute, watcher_name):
 def exponential_back_off(failures, factor, base, max_time):
     seconds = factor * base ** failures
     return seconds if seconds < max_time else max_time
+
+
+def get_service_instances_with_changed_id(marathon_client, instances, cluster):
+    marathon_app_ids = list_all_marathon_app_ids(marathon_client)
+    service_instances = []
+    for service, instance in instances:
+        config = load_marathon_service_config_no_cache(service=service,
+                                                       instance=instance,
+                                                       cluster=cluster,
+                                                       soa_dir=DEFAULT_SOA_DIR)
+        try:
+            config_app_id = config.format_marathon_app_dict()['id']
+        except NoDockerImageError:
+            config_app_id = None
+        if not config_app_id or (config_app_id not in marathon_app_ids):
+            service_instances.append((service, instance))
+    return service_instances
+
+
+def get_marathon_client_from_config():
+    marathon_config = load_marathon_config()
+    marathon_client = get_marathon_client(marathon_config.get_url(), marathon_config.get_username(),
+                                          marathon_config.get_password())
+    return marathon_client
