+++ paasta_tools/deployd/watchers.py
@@ -11,6 +11,8 @@ from kazoo.protocol.states import EventType
 from kazoo.recipe.watchers import ChildrenWatch
 from kazoo.recipe.watchers import DataWatch
 
+from paasta_tools.deployd.common import get_marathon_client_from_config
+from paasta_tools.deployd.common import get_service_instances_with_changed_id
 from paasta_tools.deployd.common import PaastaThread
 from paasta_tools.deployd.common import rate_limit_instances
 from paasta_tools.deployd.common import ServiceInstance
@@ -18,15 +20,10 @@ from paasta_tools.long_running_service_tools import AUTOSCALING_ZK_ROOT
 from paasta_tools.marathon_tools import DEFAULT_SOA_DIR
 from paasta_tools.marathon_tools import deformat_job_id
 from paasta_tools.marathon_tools import get_all_marathon_apps
-from paasta_tools.marathon_tools import get_marathon_client
-from paasta_tools.marathon_tools import list_all_marathon_app_ids
-from paasta_tools.marathon_tools import load_marathon_config
-from paasta_tools.marathon_tools import load_marathon_service_config_no_cache
 from paasta_tools.mesos_maintenance import get_draining_hosts
 from paasta_tools.utils import get_services_for_cluster
 from paasta_tools.utils import list_all_instances_for_service
 from paasta_tools.utils import load_system_paasta_config
-from paasta_tools.utils import NoDockerImageError
 from paasta_tools.utils import PATH_TO_SYSTEM_PAASTA_CONFIG_DIR
 
 
@@ -138,13 +135,6 @@ class PublicConfigFileWatcher(PaastaWatcher):
         self.is_ready = True
 
 
-def get_marathon_client_from_config():
-    marathon_config = load_marathon_config()
-    marathon_client = get_marathon_client(marathon_config.get_url(), marathon_config.get_username(),
-                                          marathon_config.get_password())
-    return marathon_client
-
-
 class MaintenanceWatcher(PaastaWatcher):
     def __init__(self, inbox_q, cluster, **kwargs):
         super(MaintenanceWatcher, self).__init__(inbox_q, cluster)
@@ -236,23 +226,6 @@ class PublicConfigEventHandler(pyinotify.ProcessEvent):
                                                          watcher_name=self.__class__.__name__)
             for service_instance in service_instances:
                 self.filewatcher.inbox_q.put(service_instance)
-
-
-def get_service_instances_with_changed_id(marathon_client, instances, cluster):
-    marathon_app_ids = list_all_marathon_app_ids(marathon_client)
-    service_instances = []
-    for service, instance in instances:
-        config = load_marathon_service_config_no_cache(service=service,
-                                                       instance=instance,
-                                                       cluster=cluster,
-                                                       soa_dir=DEFAULT_SOA_DIR)
-        try:
-            config_app_id = config.format_marathon_app_dict()['id']
-        except NoDockerImageError:
-            config_app_id = None
-        if not config_app_id or (config_app_id not in marathon_app_ids):
-            service_instances.append((service, instance))
-    return service_instances
 
 
 class YelpSoaEventHandler(pyinotify.ProcessEvent):
