+++ tests/deployd/test_common.py
@@ -6,10 +6,14 @@ import unittest
 import mock
 
 from paasta_tools.deployd.common import exponential_back_off
+from paasta_tools.deployd.common import get_marathon_client_from_config
+from paasta_tools.deployd.common import get_service_instances_with_changed_id
 from paasta_tools.deployd.common import PaastaQueue
 from paasta_tools.deployd.common import PaastaThread
 from paasta_tools.deployd.common import rate_limit_instances
 from paasta_tools.deployd.common import ServiceInstance
+from paasta_tools.utils import DEFAULT_SOA_DIR
+from paasta_tools.utils import NoDockerImageError
 
 
 class TestPaastaThread(unittest.TestCase):
@@ -62,3 +66,44 @@ def test_exponential_back_off():
     assert exponential_back_off(0, 60, 2, 6000) == 60
     assert exponential_back_off(2, 60, 2, 6000) == 240
     assert exponential_back_off(99, 60, 2, 6000) == 6000
+
+
+def test_get_service_instances_with_changed_id():
+    with mock.patch(
+        'paasta_tools.deployd.common.list_all_marathon_app_ids', autospec=True
+    ) as mock_get_marathon_apps, mock.patch(
+        'paasta_tools.deployd.common.load_marathon_service_config_no_cache', autospec=True
+    ) as mock_load_marathon_service_config:
+        mock_get_marathon_apps.return_value = ['universe.c137.c1.g1',
+                                               'universe.c138.c1.g1']
+        mock_service_instances = [('universe', 'c137'), ('universe', 'c138')]
+        mock_configs = [mock.Mock(format_marathon_app_dict=mock.Mock(return_value={'id': 'universe.c137.c1.g1'})),
+                        mock.Mock(format_marathon_app_dict=mock.Mock(return_value={'id': 'universe.c138.c2.g2'}))]
+        mock_load_marathon_service_config.side_effect = mock_configs
+        ret = get_service_instances_with_changed_id(mock.Mock(), mock_service_instances, 'westeros-prod')
+        assert mock_get_marathon_apps.called
+        calls = [mock.call(service='universe',
+                           instance='c137',
+                           cluster='westeros-prod',
+                           soa_dir=DEFAULT_SOA_DIR),
+                 mock.call(service='universe',
+                           instance='c138',
+                           cluster='westeros-prod',
+                           soa_dir=DEFAULT_SOA_DIR)]
+        mock_load_marathon_service_config.assert_has_calls(calls)
+        assert ret == [('universe', 'c138')]
+
+        mock_configs = [mock.Mock(format_marathon_app_dict=mock.Mock(side_effect=NoDockerImageError)),
+                        mock.Mock(format_marathon_app_dict=mock.Mock(return_value={'id': 'universe.c138.c2.g2'}))]
+        mock_load_marathon_service_config.side_effect = mock_configs
+        ret = get_service_instances_with_changed_id(mock.Mock(), mock_service_instances, 'westeros-prod')
+        assert ret == [('universe', 'c137'), ('universe', 'c138')]
+
+
+def test_get_marathon_client_from_config():
+    with mock.patch(
+        'paasta_tools.deployd.common.load_marathon_config', autospec=True
+    ), mock.patch(
+        'paasta_tools.deployd.common.get_marathon_client', autospec=True
+    ) as mock_marathon_client:
+        assert get_marathon_client_from_config() == mock_marathon_client.return_value
