+++ tests/deployd/test_watchers.py
@@ -8,8 +8,6 @@ import mock
 from pytest import raises
 
 from paasta_tools.deployd.common import ServiceInstance
-from paasta_tools.utils import DEFAULT_SOA_DIR
-from paasta_tools.utils import NoDockerImageError
 
 
 class FakePyinotify(object):  # pragma: no cover
@@ -248,15 +246,6 @@ class TestPublicConfigWatcher(unittest.TestCase):
         assert self.watcher.is_ready
 
 
-def test_get_marathon_client_from_config():
-    with mock.patch(
-        'paasta_tools.deployd.watchers.load_marathon_config', autospec=True
-    ), mock.patch(
-        'paasta_tools.deployd.watchers.get_marathon_client', autospec=True
-    ) as mock_marathon_client:
-        assert get_marathon_client_from_config() == mock_marathon_client.return_value
-
-
 class TestMaintenanceWatcher(unittest.TestCase):
     def setUp(self):
         self.mock_inbox_q = mock.Mock()
@@ -408,38 +397,6 @@ class TestPublicConfigEventHandler(unittest.TestCase):
             assert mock_get_service_instances_with_changed_id.called
             assert mock_rate_limit_instances.called
             self.mock_filewatcher.inbox_q.put.assert_called_with(mock_si)
-
-
-def test_get_service_instances_with_changed_id():
-    with mock.patch(
-        'paasta_tools.deployd.watchers.list_all_marathon_app_ids', autospec=True
-    ) as mock_get_marathon_apps, mock.patch(
-        'paasta_tools.deployd.watchers.load_marathon_service_config_no_cache', autospec=True
-    ) as mock_load_marathon_service_config:
-        mock_get_marathon_apps.return_value = ['universe.c137.c1.g1',
-                                               'universe.c138.c1.g1']
-        mock_service_instances = [('universe', 'c137'), ('universe', 'c138')]
-        mock_configs = [mock.Mock(format_marathon_app_dict=mock.Mock(return_value={'id': 'universe.c137.c1.g1'})),
-                        mock.Mock(format_marathon_app_dict=mock.Mock(return_value={'id': 'universe.c138.c2.g2'}))]
-        mock_load_marathon_service_config.side_effect = mock_configs
-        ret = get_service_instances_with_changed_id(mock.Mock(), mock_service_instances, 'westeros-prod')
-        assert mock_get_marathon_apps.called
-        calls = [mock.call(service='universe',
-                           instance='c137',
-                           cluster='westeros-prod',
-                           soa_dir=DEFAULT_SOA_DIR),
-                 mock.call(service='universe',
-                           instance='c138',
-                           cluster='westeros-prod',
-                           soa_dir=DEFAULT_SOA_DIR)]
-        mock_load_marathon_service_config.assert_has_calls(calls)
-        assert ret == [('universe', 'c138')]
-
-        mock_configs = [mock.Mock(format_marathon_app_dict=mock.Mock(side_effect=NoDockerImageError)),
-                        mock.Mock(format_marathon_app_dict=mock.Mock(return_value={'id': 'universe.c138.c2.g2'}))]
-        mock_load_marathon_service_config.side_effect = mock_configs
-        ret = get_service_instances_with_changed_id(mock.Mock(), mock_service_instances, 'westeros-prod')
-        assert ret == [('universe', 'c137'), ('universe', 'c138')]
 
 
 class TestYelpSoaEventHandler(unittest.TestCase):
